requirements.txt
tensorflow==2.11.0
scikit-learn==1.2.1
biopython==1.80
pandas==1.5.3
matplotlib==3.6.3
